requirements.txt
requests==2.25.0
selenium==3.14.0
Pillow==8.0.1
boto3
botocore